refactor: move InputFilesTable
