+++ src/features/AddModelDialog/InputFilesTable/InputFilesTable.styled.ts
@@ -1,6 +1,6 @@
 import { Table } from '@equinor/eds-core-react'
 import styled from 'styled-components'
-import { spacings } from '../../tokens/spacings'
+import { spacings } from '../../../tokens/spacings'
 
 const StyledTableCell = styled(Table.Cell)`
   display: flex;
